Add Vault LinkIndex for lint, link hygiene, and site backlinks.
Centralize broken-link detection and backlink indexing in vault_links.py; site build uses the same resolution rules for wikilinks and markdown page links.

Closes #52
